bump `lib-toggl` dependency to latest
That version has correct pattern to import from Pydantic v2 and fall back to v1.

Added note to dev docs to explicitly cover how to install with constraints when working locally with the dev container. Previously, was not including constraints which resulted in latest pydantic v2 being installed. Didn't run into any issues because I was not using any of the other components that still depend on v1.