fixes for issue #17 / new Time Entries not being created with tags
release v0.1.4 now depends on the updated version of `lib-toggl`
slight re-factor on `sensor.py`. Rather than use `@validator` in lib-toggl (0.1.8 removes this...) we allow both `None` and `[]` to co-exist. Only if tags and tag_ids are not empty do we bother trying to explicitly zipping tags together for attributes. In all other cases, attributes will be set to {}.

For both create and edit calls, immediately update the entity to avoid waiting for the next poll.

A few dictionary exceptions / linter fixes